refactor(lints): Pull out reporting helpers
